[refactor] move notFound component to shared, fix emptyState illustration width
